% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{algoCheck_all}
\alias{algoCheck_all}
\title{Algoritmic similarity check for both directories}
\usage{
algoCheck_all(student_dir, solution_dir)
}
\arguments{
\item{student_dir}{Object to for objective correctness checking.}

\item{solution_dir}{Object to check x against, i.e. s is the candidate solution.}
}
\value{
Dumb pasted info at this point
}
\description{
Algoritmic similarity check for both directories
}
\examples{
\dontrun{
print("Hello")
}
}
