% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{algoCheck_directory}
\alias{algoCheck_directory}
\title{Similarity Check for two libraries}
\usage{
algoCheck_directory(student_dir, solution_dir)
}
\arguments{
\item{student_dir}{Object to for objective correctness checking.}

\item{solution_dir}{Object to check x against, i.e. s is the candidate solution.}
}
\value{
Dumb pasted info at this point
}
\description{
Similarity Check for two libraries
}
\examples{
\dontrun{
algoCheck_directory(1, 1, c(0.5, 0.5))
}
}
