% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_functions.R
\name{assign_grade}
\alias{assign_grade}
\title{Function that calculates final style score}
\usage{
assign_grade(errors)
}
\arguments{
\item{errors}{Data frame of style errors for a directory of files (not a
single file)}
}
\value{
This function returns the final style scores for a directory of
scripts.
}
\description{
function that uses the data frame of errors that is returned
from style_check_file() and then multiplies it by to a weight vector that
penalizes based on how common style errors were in a directory of files (ie.
more common errors are penalized less harshly, while rare style errors are
harshly penalized). NOTE: Consider turning elements of \code{errors} into a
proportion of errors based on number of lines of code in a script. Also,
consider incorporating "consistency" of style errors across all questions
into final style score.
}
