% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.R
\name{bar}
\alias{bar}
\title{bar}
\usage{
bar(name, min = 0, max = NA)
}
\arguments{
\item{name}{character: name of variable in the data frame to display in a barchart}

\item{min}{numeric: minimal value (default: \code{0})}

\item{max}{numeric: maximal value (default: \code{NA})}
}
\value{
a list
}
\description{
Sets the name of a variable in the data frame for displaying in a barchart.
}
\examples{
bar('test')
}
