% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine.grading}
\alias{combine.grading}
\title{Combine each question script with grades/feedback and generate total
grades.csv}
\usage{
combine.grading(pset.dir)
}
\arguments{
\item{pset.dir}{directory path (scratch folder)}
}
\description{
Combine each question script, with question level grade and
feedback in "combined_submissions/". Also appends total grade at bottom of
script and creates the total grades.csv
}
