% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctness_utils.R
\name{correct_levels}
\alias{correct_levels}
\title{Strictness Levels for Objective Correctness Checking}
\usage{
correct_levels()
}
\value{
Dumb pasted info at this point
}
\description{
Strictness Levels for Objective Correctness Checking
}
