% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctness.R
\name{correctCheck_directory}
\alias{correctCheck_directory}
\title{Check the Objective Correctness of Submitted Code}
\usage{
correctCheck_directory(student_dir, solution_file, timeout = 10, ...)
}
\arguments{
\item{student_dir}{path to a single directory with submitted solution
scripts (R and not Rmd at the moment)}

\item{solution_file}{path to a solution script (R and not Rmd at the moment)}

\item{timeout}{timeout threshold.}

\item{...}{optional arguments to FUN.}
}
\value{
Dumb pasted info at this point
}
\description{
Check the Objective Correctness of Submitted Code
}
\examples{
\dontrun{
  correctCheck_directory('ClinicSubmissions-1', 'Solutions-1/Solution3.R',
                         timeout = 10)
}
}
