% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctness.R
\name{correctCheck_file}
\alias{correctCheck_file}
\title{Check the Objective Correctness of Submitted Code}
\usage{
correctCheck_file(student_file, solution_file, timeout = 10, ...)
}
\arguments{
\item{student_file}{path to a submitted script (R and not Rmd at the moment)}

\item{solution_file}{path to a solution script (R and not Rmd at the moment)}

\item{timeout}{timeout threshold.}

\item{...}{optional arguments to FUN.}
}
\description{
Check the Objective Correctness of Submitted Code
}
\examples{
\dontrun{
correctCheck_file('Submission.R', 'Solution.R', timeout = 10)
correctCheck_file('Submission.R', 'Solution.R', timeout = 10, iris)
}
}
