% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctness.R
\name{correctCheck_filesdir}
\alias{correctCheck_filesdir}
\title{Check the Objective Correctness of Submitted Code}
\usage{
correctCheck_filesdir(files, solution_file, timeout = 10, ...)
}
\arguments{
\item{files}{a character vector of file names possibly with optional paths
(and R and not Rmd at the moment)}

\item{solution_file}{path to a solution script (R and not Rmd at the moment)}

\item{timeout}{timeout threshold.}

\item{...}{optional arguments to FUN.}
}
\value{
Dumb pasted info at this point
}
\description{
Check the Objective Correctness of Submitted Code
}
\examples{
\dontrun{
  # Update this example for this function...
  correctCheck_directory('ClinicSubmissions-1',
                         'Solutions-1/Solution3.R',
                         timeout = 10)
}
}
