% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctness.R
\name{correctCheck_object}
\alias{correctCheck_object}
\title{Check the Objective Correctness of An Object}
\usage{
correctCheck_object(x, solution_file, timeout = 10, ...)
}
\arguments{
\item{x}{the object that is being compared to \code{solution_file}}

\item{solution_file}{path to a submitted solution script (R and not Rmd at
the moment)}

\item{timeout}{timeout threshold.}

\item{...}{optional arguments to FUN.}
}
\description{
Check the Objective Correctness of An Object
}
\details{
The last line of code (non-empty, non-comment line) in the solution
file is what is being evaluated. This will be an important RULE to document!
}
\examples{
\dontrun{
correctCheck_object(answer, 'Solution.R', timeout = 10)
correctCheck_object(answer, 'Solution.R', timeout = 10, iris)
}
}
