% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctness_utils.R
\name{correctTests}
\alias{correctTests}
\title{Strictness Levels for Objective Correctness Checking}
\usage{
correctTests(x, s)
}
\arguments{
\item{x}{object to for objective correctness checking.}

\item{s}{object to check x against, i.e. s is the candidate solution.}
}
\value{
dumb pasted info at this point
}
\description{
Strictness Levels for Objective Correctness Checking
}
\examples{
\dontrun{
strictness(1, 1)
}
}
