% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctness.R
\name{eval_file}
\alias{eval_file}
\title{Eval an R File}
\usage{
eval_file(filename, timeout = 10)
}
\arguments{
\item{filename}{path to a R script (R and not Rmd at the moment)}

\item{timeout}{timeout threshold.}
}
\description{
Eval an R File
}
\details{
We will need to deduce the name of the object from the
last line of code (non-empty, non-comment line) in the solution file
and this will be an important RULE to document/explain...
}
\examples{
\dontrun{
eval_file('Solution.R', timeout = 10)
}
}
