% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{feedback}
\alias{feedback}
\alias{feedback.default}
\title{feedback}
\usage{
feedback(x, ...)

\method{feedback}{default}(x, n = 30, ignore.case = TRUE, ...)
}
\arguments{
\item{x}{data frame:}

\item{...}{unused}

\item{n}{numeric: minimal subgroup size (default: \code{n=30})}

\item{ignore.case}{logical: should case differences in id variables be ignored  (default: \code{TRUE})}
}
\value{
a feedback object
}
\description{
Creates a feedback object
}
\examples{
x  <- generateTestData(n=100, 
                       IQ1=round(rnorm(100, 100, 15))\%/\%5-10,
                       IQ2=round(rnorm(100, 100, 15))\%/\%5-10,
                       course=sample(c("BWL", "VWL", "Statistik"), size=100, replace=TRUE)
                      )
fb <- feedback(x, 'ID'=id('code'), 
              'IQ Test Ergebnis 1'=bar('IQ1', min=0, max=20),
              'IQ Test Ergebnis 2'=bar('IQ2', min=0, max=20),
              'Geschlecht'=grp('sex'),
              'Studiengang'=grp('course'))
str(fb)
}
