% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.R
\name{grp}
\alias{grp}
\title{grp}
\usage{
grp(name)
}
\arguments{
\item{name}{character: name of group variable in data frame}
}
\value{
a list
}
\description{
Sets a name of group variable in the data frame.
}
\examples{
id('test')
}
