% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.R
\name{id}
\alias{id}
\title{id}
\usage{
id(name)
}
\arguments{
\item{name}{character: name of ID variable in data frame}
}
\value{
a list
}
\description{
Sets the name of ID variable in the data frame.
}
\examples{
id('test')
}
