% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userfuncs.R
\name{libraryCheck}
\alias{libraryCheck}
\title{Feedback and grading system for R coding}
\usage{
libraryCheck(student_dir, solution_dir, timeout = 10, ...)
}
\arguments{
\item{student_dir}{Path to directory of student submitted solutions}

\item{solution_dir}{Path to directory of a example solution made by instructor}

\item{timeout}{timeout threshold.}

\item{...}{Option for functions}
}
\value{
Recommended solutions library based on algorithmic, objective, style
grade as well as a database of feedback and grades for submitted solutions
}
\description{
Feedback and grading system for R coding
}
\examples{
\dontrun{
print("hello")
}
}
