% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{parse_algo}
\alias{parse_algo}
\title{Algorithm of an R File}
\usage{
parse_algo(filename)
}
\arguments{
\item{filename}{path to a R script (R and not Rmd at the moment)}
}
\description{
Algorithm of an R File
}
\details{
...
}
\examples{
\dontrun{
parse_algo('Solution.R')
}
}
