% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cleaning.R
\name{parse_sort}
\alias{parse_sort}
\title{Check if student solutions can be parsed}
\usage{
parse_sort(sol_path = ".", student_path = ".", func_wrap = FALSE)
}
\arguments{
\item{sol_path}{the path containing all solution library solutions as .R
files}

\item{student_path}{the path containing all student solutions as .R files}

\item{func_wrap}{if TRUE, will wrap all .R files in both paths in a dummy
function}
}
\value{
a list of the .R files in the student path that do not parse
}
\description{
Check if student solutions can be parsed
}
