% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ND_funcs_1202.R
\name{pset.skeleton}
\alias{pset.skeleton}
\title{Create a skeleton for a new problem set in a folder that does not live
in the QLibrary folder}
\usage{
pset.skeleton(dir, qdirs)
}
\arguments{
\item{dir}{directory path for a new problem set (scratch folder)}

\item{qdirs}{vector of directory paths to each question that the instructor
wants to use for this problem set, ordered appropriately}
}
\description{
function creates these blank files and folders:
\itemize{
\item header.R - a header and an introduction to the problem set
\item stylemethod.R - instructions to feedback on how to evaluate style
\item other.R (TBD)
\item pset_body.R - the assembled questions in blocks
\item q_names.csv - a csv file that lines up with pset_body.R, giving the
proper order, question names (unique at the level of this problem set),
and paths to the question folders in QLibrary.
\item Make copies of the QLibrary question folders in this specified
scratch folder, dir, and only at the very very end, and very carefully,
would we "push back" to the QLibrary folders.
\item ?SKIP? Make a pset-level data/ folder, with all data from the
question folders
}
}
