% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ND_funcs_1202.R
\name{question.skeleton}
\alias{question.skeleton}
\title{Create a skeleton for a new question in the QLibrary folder}
\usage{
question.skeleton(dir = NA, qfile = NA)
}
\arguments{
\item{dir}{path to a directory where skeleton for a new question will be
added to the question library (QLibrary)}

\item{qfile}{path to a script that follows the "rules" and is used for
the statement.R file.  This is optional.}
}
\description{
function creates these blank files and folders:
statement.R - the question which a student shall solve
solutions/ - folder to store the solution library
solutions/mastersolution.R - instructor solution
submission_archive/ - folder to store all submitted student scripts
data/ - if the question involves a data set it shall be stored here
correctmethod.R (optional, TBD for now)
algomethod.R (optional, TBD for now)
stylemethod.R (for desired style formatting)
}
