% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_grading.R
\name{start_grading}
\alias{start_grading}
\title{Function that splits solution and submissions into question-level
folders. It does the style, correctness and similarity checks using
\code{stylemethod.R}, \code{correctmethod.R}, and \code{algomethod.R}, generating the
corresponding scores.}
\usage{
start_grading(dir)
}
\arguments{
\item{dir}{directory path for a new problem set (scratch folder)}
}
\description{
This function incorporates splitting the problem set and grading
it. It splits the mastersolution.R and all submissions given in \code{dir} into
question-level folders (ie. grades-Q1/, sol_dir-Q1/, submission-Q1/), and
places the split scripts into their correspoding folders. It does style check
and correctness check at the question level. For style check, it also appends
written feedback (not scores) at the end of question-level submissions. It
then records the grades in question-level gradebook CSVs. After this is done,
questions can be passed off to TFs for the algo work.
}
