% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{style.80char}
\alias{style.80char}
\title{Provide 80-character width feedback}
\usage{
style.80char(submissionfolder)
}
\arguments{
\item{submissionfolder}{a character string describing the directory of scripts
to be checked}
}
\value{
Need to define
}
\description{
Provide 80-character width feedback
}
