% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{style.comment}
\alias{style.comment}
\title{Provide comment feedback}
\usage{
style.comment(submissionfolder)
}
\arguments{
\item{submissionfolder}{a character string describing the directory of scripts
to be checked}
}
\value{
Need to define
}
\description{
Provide comment feedback
}
