% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_functions.R
\name{style_check_dir}
\alias{style_check_dir}
\title{Function checks style on a directory of scripts (not individual files).}
\usage{
style_check_dir(sub_folder, linter = available_linters())
}
\arguments{
\item{sub_folder}{directory path on which to check style}

\item{linter}{string of linter names or tags used to check style. This is
provided by style_input(). The default is to use all linters found in
available_linters(). NOTE: Consider creating a csv of desired linters as
documented in lint(). Current implementation runs lint() using all linters,
and we only select the desired ones.}
}
\value{
This function returns a data frame containing the number of errors in
all scripts for each inputted lintr. These errors will be later used with a
weight vector to calculate the final style score.
}
\description{
function that uses lintr to check the style format of a
directory of scripts. \code{linter} is obtained from instructor by previously
using style_input().
}
