% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_functions.R
\name{style_check_file}
\alias{style_check_file}
\title{Function checks style of individual files (not a directory of files).}
\usage{
style_check_file(filename, linterList = NULL, linterTags = NULL)
}
\arguments{
\item{filename}{script path on which to check style}

\item{linterList}{string of linter names (not tags) used to check style}

\item{linterTags}{single linter tag (not specific linter names) used to check
style}
}
\value{
This function returns the number of errors in the script for
each inputted lintr. These errors will be later used with a weight vector to
calculate the final style score.
}
\description{
function that uses lintr to check the style format of a single
script. The parameters are obtained from instructor by previously using
style_input(). This function returns the number of errors in the script for
each inputted lintr. These errors will be later used with a weight vector to
calculate the final style score.
NOTE: Use linterList or linterTags, not both at same time. Consider changing
fucntion to only take one parameter for linters, regardless of whether linter
names or tags.
}
