% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_functions.R
\name{style_feedback}
\alias{style_feedback}
\title{Function returns written style feedback for a single script}
\usage{
style_feedback(filename, linter = available_linters())
}
\arguments{
\item{filename}{script path used to generate style feedback}

\item{linter}{string of linter names or tags used to check style. This is
provided by style_input(). The default is to use all linters found in
available_linters(). NOTE: Consider creating a csv of desired linters as
documented in lint(). Current implementation runs lint() using all linters,
and we only select the desired ones.
style}
}
\value{
This function returns a data frame with the first instance of
each style error, along with the line number of which it occurred and a
message on how to properly style that code.
}
\description{
function that uses lintr to check the style format of a single
script. The parameters are obtained from instructor by previously using
style_input(). This function returns a data frame with the first instance of
each style error, along with the line number of which it occurred and a
message on how to properly style that code.
}
