% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stylemethod.R
\name{style_input}
\alias{style_input}
\title{Instructor inputs which linters to later use in style_check_file()}
\usage{
style_input()
}
\value{
Returns the inputted linters/tags to then be used in style_check_file().
}
\description{
This function displays the available linters (or linter tags) to
the instructor and prompts the instructor to input which linters to use. It
returns the inputted linters/tags to then be used in style_check_file().
NOTE: Consider whether this function can be completely moved into
style_check_file(). Parameters are TBD. Right now, no parameters
}
