% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{styleCheck_directory}
\alias{styleCheck_directory}
\title{Check the Style of Submitted Code}
\usage{
styleCheck_directory(dirname)
}
\arguments{
\item{dirname}{path to a directory with submitted solution scripts (R
and not Rmd at the moment)}
}
\value{
Dumb 10 at this point
}
\description{
Check the Style of Submitted Code
}
\details{
Just a wrapper so check everything in a directory, although
we should think about collection of results.
}
\examples{
styleCheck_directory("Hello")
}
