% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{styleCheck_file}
\alias{styleCheck_file}
\title{Check the Style of Submitted Code}
\usage{
styleCheck_file(filename, verbose = FALSE)
}
\arguments{
\item{filename}{the name of a file with a path}

\item{verbose}{debugging temporary option?}
}
\description{
Check the Style of Submitted Code
}
\details{
Wait a sec, my foo.R example in the tests has bad indentation?
Apparently \code{lintr} isn't detecting indentation, while \code{styler} and
\code{formatR} both do (though clunkier to work with).  The algebra here right
now is just placeholder work showing the starting point for the objects
to be studied/evaluated.  Need to think about collection of results
for various parts of this project, actually.  Because \code{lintr} only seems
to work on a file, I don't envision a style checking on an objects, so if
we have an object it would require creating a temp file and then using
this function.  Feels a little backwards, but hey.
}
