% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.feedback.R
\name{summary.feedback}
\alias{summary.feedback}
\title{summary.feedback}
\usage{
\method{summary}{feedback}(object, ...)
}
\arguments{
\item{object}{feedback object}

\item{...}{ignored}
}
\value{
nothing
}
\description{
Prints a summary of a feedback object
}
\examples{
x  <- generateTestData(n=100, 
                       IQ1=round(rnorm(100, 100, 15))\%/\%5-10,
                       IQ2=round(rnorm(100, 100, 15))\%/\%5-10,
                       course=sample(c("BWL", "VWL", "Statistik"), size=100, replace=TRUE)
                      )
fb <- feedback(x, 'ID'=id('code'), 
              'IQ Test Ergebnis 1'=bar('IQ1', min=0, max=20),
              'IQ Test Ergebnis 2'=bar('IQ2', min=0, max=20),
              'Geschlecht'=grp('sex'),
              'Studiengang'=grp('course'))
summary(fb)
}
