% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctness.R
\name{timeoutCatch}
\alias{timeoutCatch}
\title{Try Catch for internal use}
\usage{
timeoutCatch(func, timeout = 10, ...)
}
\arguments{
\item{func}{the function to be evaluated.}

\item{timeout}{timeout threshold.}

\item{...}{optional arguments to FUN.}
}
\description{
Try Catch for internal use
}
\examples{
\dontrun{
timeoutCatch(myfunction, timeout = 10, iris)
}
}
